+++ SpaceTraders/app/src/main/java/cs2340/spacetraders/model/Model.java
@@ -1,0 +1,4 @@
+package cs2340.spacetraders.model;
+
+public class Model {
+}
