+++ SpaceTraders/app/src/main/java/cs2340/spacetraders/viewmodels/ConfigurationViewModel.java
@@ -1,0 +1,4 @@
+package cs2340.spacetraders.viewmodels;
+
+public class ConfigurationViewModel {
+}
